+++ 4get/4get/logic_task_deadline.cpp
@@ -4,14 +4,14 @@ TaskDeadline::TaskDeadline()
 {
 }
 
-TaskDeadline::TaskDeadline(int id, string description, string location, tm *reminderTime, Priority priority, Status status, RepeatType repeat, tm *endTime) : Task()
+TaskDeadline::TaskDeadline(long long id, string description, string location, tm *reminderTime, Priority priority, Status status, RepeatType repeat, tm *endTime) : Task()
 {
 	TaskType type = deadline;
 	setupTask(id, type, description, location, reminderTime, priority, status, repeat, endTime);
 }
 
 
-void TaskDeadline::setupTask(int id, TaskType type, string description, string location, tm *reminderTime, Priority priority, Status status, RepeatType repeat, tm *endTime)
+void TaskDeadline::setupTask(long long id, TaskType type, string description, string location, tm *reminderTime, Priority priority, Status status, RepeatType repeat, tm *endTime)
 {
 	taskId = id;
 	taskType = type;
