+++ 4get/4get/logic_task_deadline.h
@@ -8,8 +8,8 @@ class TaskDeadline : public Task
 {
 public:
 	TaskDeadline();
-	TaskDeadline(int id, string description, string location, tm* reminderTime, Priority priority, Status status, RepeatType repeat, tm* endTime);
-	virtual void setupTask(int id, TaskType type, string description, string location, tm *reminderTime, Priority priority, Status status, RepeatType repeat, tm *endTime);
+	TaskDeadline(long long id, string description, string location, tm* reminderTime, Priority priority, Status status, RepeatType repeat, tm* endTime);
+	virtual void setupTask(long long id, TaskType type, string description, string location, tm *reminderTime, Priority priority, Status status, RepeatType repeat, tm *endTime);
 	void setTaskEnd(tm *endTask);
 	tm* getTaskEnd();
 };
